fix(download): graceful shutdown and proper interrupt handling
The download pipeline had several bugs that caused progress loss on
pause/close and corrupted state on interrupted files.

service/files.rs:
- Introduce DownloadOutcome enum (Completed/Interrupted). A cancelled
  download no longer returns Ok(()) so workers do not mark partial
  files as fully downloaded and do not forward them to the unzip stage.
- Persist .part sidecar every SAVE_INTERVAL and on interrupt.

start_download_version.rs / continue_download_version.rs:
- Add MAX_DOWNLOAD_RETRIES=5 instead of an infinite retry loop.
- Handle DownloadOutcome::Interrupted: persist partial size, stop
  without counting the file as completed.
- Add persist_file_size() helper writing FileProgress.size from .part.
- Move unzip config update out of spawn_blocking (no more block_on
  inside a blocking thread — eliminates pool starvation risk).
- Fix double increment of downloaded_cnt in continue's to_unpack block.

handlers/window.rs + lib.rs:
- Replace raw process::exit(0) with graceful_shutdown(): signal all
  active downloads to stop, wait for workers to flush .part files and
  config, then save.
- Add on_window_event(CloseRequested) hook so closing the window on
  the X button also goes through graceful shutdown.
- Add cancel_all_downloads_and_save command.

+++ src-tauri/src/handlers/dto.rs
@@ -29,9 +29,9 @@ pub struct DownloadProgress {
   #[serde(default)]
   pub file: String,
   #[serde(default)]
-  pub downloaded_files_cnt: u16,
+  pub downloaded_files_cnt: u32,
   #[serde(default)]
-  pub total_file_count: u16,
+  pub total_file_count: u32,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
